# Network Hierarchy optimizations for rebuilding hierarchies