**Before**

```
Benchmark                                                                     Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------
ServerDeepHierarchyBenchmark/RebuildHierarchy                              43.7 us         43.9 us        16000
ServerDeepHierarchyBenchmark/RebuildHierarchyRemoveAndAddLastChild          133 us          134 us         5600
ServerDeepHierarchyBenchmark/RebuildHierarchyRemoveAndAddMiddleChild        167 us          165 us         4073
ServerDeepHierarchyBenchmark/RebuildHierarchyRemoveAndAddFirstChild         231 us          225 us         2987
```

**After**

```
Benchmark                                                                     Time             CPU   Iterations
---------------------------------------------------------------------------------------------------------------
ServerDeepHierarchyBenchmark/RebuildHierarchy                              30.5 us         30.7 us        22400
ServerDeepHierarchyBenchmark/RebuildHierarchyRemoveAndAddLastChild         67.9 us         68.0 us         8960
ServerDeepHierarchyBenchmark/RebuildHierarchyRemoveAndAddMiddleChild       70.6 us         69.8 us        11200
ServerDeepHierarchyBenchmark/RebuildHierarchyRemoveAndAddFirstChild        72.0 us         73.2 us         8960
```

+ reworked recursive rebuilding to iterative breadth first method
+ some minor optimization here and there

Note: the biggest item now is TransformBus::GetChildren

![image](https://user-images.githubusercontent.com/5432499/136090908-358c92d5-1a09-40b0-af04-3daab07212a0.png)